# feat: Speed up description parsing, support use of comments and other…