@hermanschaaf 2022-08-29
Closing this for now, as we might not support description parsing going forward.